update kafka requirement to latest (#153)
Thanks to @matdehaast for the contribution.

Updates the Kafka requirement to the latest version/ Fixes bugs of values not being pulled through correctly for the Zookeeper child. The replica count for the Zookeeper will now be applied correctly. The benefit being that the CPU usage will be reduced for deployments where high-availability is not necessary.